V1.0

+++ lib/general/include/bitbuffer.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <stdint.h>
 
-uint32_t make_bit_buffer(volatile uint32_t* buffer, const uint8_t* pulse_length);
+uint32_t make_bit_buffer(volatile uint32_t* buffer, const uint16_t* pulse_length);
